Apply cargo clippy --fix auto-fixes
Automatically fixed 215+ clippy warnings using cargo clippy --fix.
Reduced warnings from 284 to 69.

Changes include:
- Removed unnecessary references
- Fixed floating point literal formatting
- Other clippy suggestions

+++ src/data/move_callbacks/dive.rs
@@ -159,7 +159,7 @@ pub mod condition {
         // }
         if move_id == "surf" || move_id == "whirlpool" {
             // return this.chainModify(2);
-            return EventResult::Number(battle.chain_modify(2.0 as f32));
+            return EventResult::Number(battle.chain_modify(2.0_f32));
         }
 
         EventResult::Continue
