+++ src/data/move_callbacks/facade.rs
@@ -28,7 +28,7 @@ pub fn on_base_power(battle: &mut Battle, _base_power: i32, pokemon_pos: (usize,
 
     if status != ID::from("") && status != ID::from("slp") {
         // return this.chainModify(2);
-        return EventResult::Number(battle.chain_modify(2.0 as f32));
+        return EventResult::Number(battle.chain_modify(2.0_f32));
     }
 
     EventResult::Continue
