+++ src/data/move_callbacks/ficklebeam.rs
@@ -34,7 +34,7 @@ pub fn on_base_power(battle: &mut Battle, _base_power: i32, pokemon_pos: (usize,
         battle.add("-activate", &[pokemon_ident.as_str().into(), "move: Fickle Beam".into()]);
 
         // return this.chainModify(2);
-        return EventResult::Number(battle.chain_modify(2.0 as f32));
+        return EventResult::Number(battle.chain_modify(2.0_f32));
     }
 
     EventResult::Continue
