+++ src/data/move_callbacks/guardianofalola.rs
@@ -32,7 +32,7 @@ pub fn damage_callback(battle: &mut Battle, _pokemon_pos: (usize, usize), target
             Some(p) => p,
             None => return EventResult::Continue,
         };
-        (target_pokemon.get_undynamaxed_hp() * 3 / 4) as i32
+        target_pokemon.get_undynamaxed_hp() * 3 / 4
     };
 
     // if (target.volatiles['protect'] || target.volatiles['banefulbunker'] || target.volatiles['kingsshield'] ||
