+++ src/data/move_callbacks/minimize.rs
@@ -42,7 +42,7 @@ pub mod condition {
         // if (boostedMoves.includes(move.id)) {
         if boosted_moves.contains(&current_move_id) {
             // return this.chainModify(2);
-            return EventResult::Number(battle.chain_modify(2 as f32));
+            return EventResult::Number(battle.chain_modify(2_f32));
         }
 
         EventResult::Continue
