+++ src/data/move_callbacks/retaliate.rs
@@ -30,7 +30,7 @@ pub fn on_base_power(battle: &mut Battle, _base_power: i32, pokemon_pos: (usize,
 
     if fainted_last_turn.is_some() {
         battle.debug("Boosted for a faint last turn");
-        return EventResult::Number(battle.chain_modify(2 as f32));
+        return EventResult::Number(battle.chain_modify(2_f32));
     }
 
     EventResult::Continue
