+++ src/data/move_callbacks/shoreup.rs
@@ -56,7 +56,7 @@ pub fn on_hit(battle: &mut Battle, pokemon_pos: (usize, usize), _target_pos: Opt
         pokemon_pokemon.maxhp
     };
 
-    let heal_amount = battle.modify_f(maxhp, factor as f64);
+    let heal_amount = battle.modify_f(maxhp, factor);
     let heal_result = battle.heal(heal_amount, Some(pokemon), None, None);
     let success = heal_result != Some(0) && heal_result.is_some();
